+++ INTEGRA-backend/app/Http/Controllers/MarketingControllers/TvController.php
@@ -14,22 +14,11 @@ class TvController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index() : TvCollection
     {
         return new TvCollection(Tv::all());
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
 
@@ -54,31 +43,24 @@ class TvController extends Controller
             'advertising_period' => request('advertising_period') ,
             'expected_revenue'   => request('expected_revenue') ,
             'campaign_id'        => request('campaign_id')
-        ]));
+        ]))
 
-        return response()->json(['message' => 'Tv has been created']);
+            return $this->success();
+        else
+            return $this->failure();    
+    
     }
 
-
-    /**
-     * Display the specified resource.
-     */
     public function show($id)
     {
-        return new TvResource(Tv::find($id));
+        $tv = Tv::find($id);
+
+        if($tv)
+            return new Tv($lead);
+        else
+            return $this->failure(); 
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request,  $id)
     {
 
@@ -98,31 +80,29 @@ class TvController extends Controller
 
         $tv = Tv::findOrFail($id);
 
-        $tv->channel                        = request('channel');
-        $tv->time                           = request('time');
-        $tv->cost                           = request('cost');
-        $tv->advertising_period             = request('advertising_period');
-        $tv->expected_revenue      = request('expected_revenue');
-        $tv->campaign_id                    = request('campaign_id');
+        $tv->channel            = request('channel');
+        $tv->time               = request('time');
+        $tv->cost               = request('cost');
+        $tv->advertising_period = request('advertising_period');
+        $tv->expected_revenue   = request('expected_revenue');
+        $tv->campaign_id        = request('campaign_id');
 
         if($tv->isDirty(['channel' , 'time' , 'cost', 'advertising_period' , 'expected_revenue' , 'campaign_id' ])){
             $tv->save();
-            return response()->json( ['message' => 'Tv is Updated']);
+            return $this->success();
         }
-
-        else {
-            return response()->json( ['message' => 'Nothing changed']);
-        }
+        else 
+            return $this->failure(); 
+        
     }
 
-
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy($id)
     {
-        $tv = Tv::findOrFail($id);
-        $tv->delete();
-        return response()->json( ['message' => 'tv has been deleted']);
+        if( $tv = Tv::findOrFail($id)) {
+            $tv->delete();
+            return $this->success();
+        } 
+        else
+            return $this->failure();
     }
 }
